Move fdtransfer server & shared files to src/fdtransfer
